+++ domain/src/main/java/app/ch/currencyconverter/domain/quote/repository/QuoteRepository.kt
@@ -2,8 +2,9 @@ package app.ch.currencyconverter.domain.quote.repository
 
 import app.ch.currencyconverter.domain.base.Error
 import app.ch.currencyconverter.domain.base.Result
+import java.math.BigDecimal
 
 interface QuoteRepository {
 
-    suspend fun getCurrencyQuotes(): Result<Map<String, Double>, Error>
+    suspend fun getCurrencyQuotes(): Result<Map<String, BigDecimal>, Error>
 }
